requirements.txt
flask==3.0.0
google-generativeai==0.8.3
python-docx==1.1.0
python-dotenv==1.0.0
scikit-learn==1.3.2
numpy==1.26.2
gunicorn==21.2.0
